Fix return value
JerryScript-DCO-1.0-Signed-off-by: László Langó llango.u-szeged@partner.samsung.com

+++ src/example1.c
@@ -27,5 +27,5 @@ main (int argc, char * argv[])
                                                           script_size,
                                                           JERRY_FLAG_EMPTY);
 
-  return return_code;
+  return (int) return_code;
 }
